Can now perform some linear regression

+++ Regression_CSharp/BacteriaRegression/BusinessLogic/BacteriaImporter.cs
@@ -22,7 +22,7 @@ namespace BacteriaRegression.BusinessLogic
                     ASV = columns[0],
                     Genus = columns[1].Split('_').Last(),
                     Class = columns[2].Split('_').Last(),
-                    Measurements = columns.Skip(3).Select(measurement => float.Parse(measurement)).ToList()
+                    Measurements = columns.Skip(3).Select(measurement => double.Parse(measurement)).ToList()
                 })
                 .ToList();
         }
